Changed MessageBox
I dropped the functionality of the expand-to-show-form for the
MessageBox. Now it's just MessageButton, which when clicked opens the
user's email client. I don't think the user really needs to use my site
to actually send the email, and would rather focus on other stuff.

One thing to take away from how the MessageBox was, is the animations.
I do still want to animate other elements of the UI.