Implemented project data, and basic list scrolling
Any component with the projectlist.json import can access the data
for the projects. Right now the only 3 that should access it are
FloatingList, ProjectList, and ProjectDetail. I'm doing it this way
instead of sending one copy around as props because it's not a very
complex set of data, and not all the data is used by each Component.

Also implemented a basic horizontal scrolling that relies on a proper
calculated width based on how many projects. I do need to adjust the
horizontal scrolling so it only scrolls horizontally via normal mouse
scrolling on the Projects view. The Project Detail view should scroll
vertically like normal.