Reworked ProjectDetail and ProjectList
Now the FloatingList appears on both the Projects view and Detail view,
but act differently according to the location. On the Detail view it
acts as a redirect jump table and takes you to that detail page.

Still need to add a switch statement into the render method for Content
so I can render a different content component depending on the project.

Also note that there's a bug that popped up: if you click on a detail
page and then use the FloatingList to redirect to a new detail page,
then click on `projects` link, the Projects view does not scroll
horizontally using the mouse as it usually does. This is likely due to
the listener that's established on window.onload isn't working properly
due to the redirects. Not sure why, but might be able to fix this by
just changing when that actionListener is attached.