Setup dynamic content for ProjectDetail pages
The Content component switches on the project id to choose the right
custom layout for that project. In the projectlist.json, each project
should have an images array that is a list of its images. Soon, each
will also have a copy array, which will be an array of copy (I think I
will try to do it this way).