Fixed some of the errors on Project Detail pages
I had set up the first project with a slightly altered system of
referencing it's respective images, and hadn't set that up for the other
projects in `projectlist.json`. Now it has placeholders at least.

Biggest thing I'm looking into fixing now is the horizontal scrolling,
to make it work properly on the Project List view and not interfere with
normal scrolling on the Project Detail views.