Fixed #9 – added Floating List anchor animation
On the Project List view, when you click on a project from the jump
table, the user is smoothly auto-scrolled to that project in the list.

Of course, this is only on the Project List view, and not the Project
Detail (the list acts as page links on that view).