Started rebuilding Project Detail view
The detail view needs to be redesigned to accommodate a new layout.
Above the fold we'll have the "Brief" which includes the project's
title, tags, summary, and preview image.

Below the fold there will be a horizontally scrolling gallery of images
for the project (some text too).

The scroll behavior needs to be controlled. When scrolling down from the
Brief, it will snap to fill the screen with the Gallery. Once you're
in the gallery, it should keep you scrolling left and right until left
is 0 (the gallery is reset to the beginning). Then it will allow you to
scroll up again. Alternatively, click the scroll-up button at any time
to scroll back up. If you scroll back up, the horizontal gallery should
not be reset to the beginning if the user scrolls back down.

To achieve that behavior, I may need to move the current scroll behavior
code to the parent DetailWrapper component, as its currently in Brief.
If I want the entire screen to be scrolled back up dependent on the
Gallery, I think I will Need it to be handled by the parent. Or,
actually I could just use `element.scrollIntoView` if left is 0, and
just get that element using `getElementById`. I think it just depends
on whether or not I want to centralize all scroll behavior or keep it
relative to the respective components.