Worked on Detail redesign and new scroll behavior
The biggest design change for the Detail view was the horizontally
scrolling bottom half, and only vertical scrolling top half (like an L).
I've written most of the code to achieve the scrollY motion, using the
mousewheel and the scroll button (I might need to add behavior control
for the actual scroll bar if a user tries to move it manually).

What still needs to be done:
- Gallery layout (its a placeholder right now)
  - Horizontal scrolling
  - Modular CSS styling for building blocks (to make building each page
    relatively quick if possible)
- Finish building all of one project (Spectra or OS1).
- Create code templates out of the current Brief and Gallery components
- Maybe handle horizontal scrolling behavior when in the Brief section
  (basically just block it completely, but that seems kind of user
  hostile. Maybe do a bouncy thing like when you're at the edge of a
  scrolling area in macOS).