Finished adjusting scroll behavior for Detail view
Now the scrolling is controlled for both vertical and horizontal
scrolling. Now all that's next for the Detail page redesign (#13) is to
build some css styles for the gallery section.

I'm thinking a S/M/L size for images, and then a small and large text
box for the copy. It doesn't need to be completely modular, it might get
some individual project styling, but I want something to start working
with to relatively quickly begin building the other pages.