Fixed bug with Detail scrolling behavior
In the middle of the scroll-up behavior (from the Gallery section), if
the wheel event changed direction, or if there were a large amount of
wheel events in general, the scroll animation would be cut short and the
user would be floating around between the Brief and Gallery section, and
had to reset by using the scroll button.

This has been fixed. In the Gallery secton's handleScroll method, a
check to ignore scroll events if it was in-between the state of the top
of the page or the top of the gallery section was preventing it from
hitting the `e.preventDefault()` call at the bottom of the function. So
it was ignoring My code, but not ignoring default scroll behavior. Duh!

Seems to be a likely fix for both trackpad and mousewheel wheel events,
but only tested on trackpad.