Attempted to fix residual bug mentioned in f63e80d
See f63e80d for details on the bug, but I think its also related
to async code. I moved `e.preventDefault()` to the top of the function
to hopefully avoid that. It was previously right after a function that
calls `setState` which I know is async. I had never noticed this
bug until the day I added the `setState` call to the code, so I'm pretty
sure that's the culprit somehow. Either way, keep an eye out.

It likely won't happen under normal usage, but I've been trying to break
it with mousewheel and trackpad, and for now it seems okay.