Started building out Spectra detail
I'm figuring out the timeline nav for jumping to different years and
indicating the current content's year. Right now it's always visible, so
that needs to be toggled based on scroll-behavior toggles (when Gallery
section is in view). Also think about dropping the line; it might look
pretty good with just the years.

I can reuse the smooth scroll code from the Project List view to handle
jumping to a specific year-start. I will need some JS code to detect
when the user has scrolled to a new year-start and set the timeline nav
to indicate the current year.