Fixed #20 – navbar stays in place
I abstracted the code that handles adjusting the navbar bottom margin
for pages with a horizontal scrollbar into `tools.js`. I can import and
call the function in any parent component that would have this issue.

I tried putting it in the parent-most container, but I realized that
Main actually does not include the Navbar component in it. I didn't
want to add this code to App. It living in tools makes sense to me, and
doesn't beef up lifecycle methods much in the relevant components so I
think it's all good. Also minimizes checks that would need to happen
every time a route changes, which might be harder to detect. (The naive
way would be to check on every update, but that's just silly!)