Fixed #19 – text doesn't overflow columns anymore
I also fixed a potential future issue where a column has ONLY a p tag,
and the text would overflow until it ends (wouldn't wrap). It's however
hardcoded to a max-width if the function detects only a p tag. Being
able to change that with a css id would be a solution for specific
bodies of text I want to take up more of the screen.