Fixed #26 – updated time link behavior
I cleared up the issue with the timeline links not working in a logical
way. Now they stay active when you're in that period, rather than
only when you hit the beginning of the period.

Also tweaked some css class names for readability and making the
paired timeline LINKs and timeline MARKERS/PERIODS relationship a little
more obvious.