Fixed #25 (and something I broke in fixing #26)
Error is gone, I was right; I just had the `onClick` func set on the
list element instead of the anchor. The behavior was just trickling down
to the child (anchor) the way I initially wrote it.

In #26 I just forgot to change the querySelector that handles changing
the links on click to use the new css class name. ez pz.