Working on Page Animation
I got it working with just CSS and JS but only for the main pages.

Now I'm trying to make it work for switching projects from the Detail
view. My attempt was to check if Brief/Gallery were being updated (i.e.
switching project content) and running the same animation. That isn't
working perfectly. The animation starts but hangs until images are
loaded. Dunno why!!!!

So I tried fixing that with a delay in the animation (didn't work) and
then with setTimeout (works but with Nav flicker).