Fixed #39 and #35 – prevents some scrolling bugs
All I did was add an event listener for all wheel events when the Detail
view is mounted (and removed on unmount). It simply prevents the default
event behavior, since I'm handling all handling specifically for each
section of the Detail view.

I can do a similar thing for #17 when I figure out how I want to handle
scroll events (left and right) when the Brief section is in view.