Fixed #39 – scroll behavior fixed
All I did was add a conditional to check if the currentView was
BRIEF, if so then block horizontal scrolling behavior. I saw it fire
on scroll-up while the mouse is hovering over the Gallery div still in
view.

Previously if you engaged the wheel event while the mouse was hovering
over the Gallery section, WHILE the scroll-up animation was playing,
you'd start scrolling the screen horizontally and the scroll-up
animation would stop. You'd be stuck in no man's land. This fix seems
like even when async would allow it to happen, it will get blocked.