Fixed #37 – Gallery shift handled onfocus
I used a focus event listener to detect when the page was being returned
to after leaving it. This code will readjust the scroll to the bottom
of the page if it meets certain conditions. The bug was only occurring
when the Gallery section was in view, and the user was returning to
the page from a Chrome New Tab tab. The Bookmark Bar added clientHeight
so it reduced the innerHeight/window's viewport height. This also messed
with the document's scrollHeight. Great!

So I had to manage all that, and really it only fires now if you're
returning and you're in the Gallery section. That means its running
every window focus event, however I wouldn't imagine that its much of a
performance cost to scroll to the position it's already at (which is
what technically happens every time it regains focus after being on any
other tab).

I also added a resize event listener to help manage a new state that is
used by the handleFocus function. I might be able to cut this out if I
rethink the handleFocus func a bit with it's conditional statements, but
for now I'll leave it. It works!