Fixed #44 – Added dynamic GalleryNav links
If a _Gallery component has a GalleryNav, it should also have a
galleryNavLinks prop this is set in defaultProps.

Check /spectra/_Gallery.js for an example. Basically it just passes an
array of link objects to the GalleryNav, and the GalleryNav constructor
creates the li and a tags (similarly to how ProjectList handles the
projects state).