Figured out logic for updating gradient
I can replace the placeholder colors, but now I am keeping track of the
previous mouse region, and using that to set the color1 and color2.
color1 is the current region, color2 is the previous region. That way
the color is always THIS color to THAT color, and THAT color is where
you came from, so hopefully you'll recognize a trail of interaction.

At least that's the idea.