Worked on #70 – new ProjectList styles for mobile
I'm not sure if it's DONE, but I added styles for the mobile nav to
be more visible for smaller screens. Now it's kinda of like an app nav
bar at the bottom. I've only tested it on the ProjectList view but I
imagine there shouldn't be issues on the others.

I also redesigned the mobile experience for the project list. Instead of
keeping it horizontal, I went with vertical so you can see multiple on
one screen. Another option would be a grid, but tbh I think how I have
it offers a decent scrolling experience.

The floating list disappears on smaller screens and honestly I don't
think it needs a mobile alternative. The mobile usage will be limited
because the mobile audience is going to be tiny.