Worked on #86
For the Info page, I added responsive styling for the screen height.

I should probably give the other pages similar treatment, which just
include using the same height breakpoint (732px), but adjust top-margins
and text size to prevent it from breaking into the nav menu (which is
the main problem with this issue).