Working on #89 – Safari and FF Detail fix (broken)
The changes in this cm offer a still pretty broken implementation
of "smooth scrolling", really just an alternative way to handle the
scroll behavior for going between Brief and Gallery sections of the
Detail view.

Using `jump.js` but that doesn't really solve the problem, and
introduces a new problem of the scroll event firing jump's function
every event (so I had to use a debounce wrapper but its still goofy
sometimes).

Ugh!