Worked on #89 – good progress for a fix
This fix works in Chrome and Safari, but its only been tested with
a mousewheel so far.

This fix does not work for Firefox. The next update will be to only
utilize this fix for Safari, and use vanilla scroll func for Chrome
and FF (I'm pretty sure it works fine there).