Fixed #91 – removed zenscroll anchor behavior
Zenscroll was automatically handling anchor links; I didn't want that,
because it was directly affecting the behaviors I set up for the
gallery nav links.

I also realized that I need to add zenscroll to the arrow buttons for
Safari, in addition to still needing to add the horizontal scrolling
fix.