Tweaked scrolling for #93
Now on scroll events, zenscroll is used for all browsers. I think I got
it to look similar to the natural smooth scrolling based on the
position I'm telling it to scroll to (it goes faster if I tell it to
scroll by a high amount, even if it moves the same distance), and the
duration.

I didn't change the logic for the buttons, maybe I'll do this later, but
the buttons never had any issues and I don't really care at this point.

Seems all good, but need to do a final test with trackpad again.